+++ test/e2e/index.spec.js
@@ -68,6 +68,7 @@ describe('evrythng.js', () => {
     require('./entity/roles.spec')('operator')
     require('./entity/rules.spec')()
     require('./entity/secretKey.spec')()
+    require('./entity/shipmentNotice.spec')()
     require('./entity/shortDomains.spec')()
     require('./entity/tasks.spec')()
     require('./entity/thngs.spec')('operator')